[refactor]: Refactoring of rating controller
