+++ Team54/src/main/java/ISA/Team54/rating/mapper/RatingMapper.java
@@ -1,12 +1,23 @@
 package ISA.Team54.rating.mapper;
 
+import ISA.Team54.rating.dto.RatingDTO;
 import ISA.Team54.rating.dto.RatingRequestDTO;
 import ISA.Team54.rating.model.Rating;
 
 public class RatingMapper {
-    public Rating RatingToRatingDTO(Rating rating){
-        return new Rating(
+    public RatingDTO RatingToRatingDTO(Rating rating){
+        RatingDTO ratingDTO = new RatingDTO();
+        ratingDTO.setId(rating.getId());
+        ratingDTO.setRating(rating.getRating());
+        ratingDTO.setRatingObject(getTypeFromRating(rating));
+        return ratingDTO;
+    }
 
-        );
+    private String getTypeFromRating(Rating rating) {
+        if(rating.getDermatologist() != null)
+            return "Dermatolog";
+        else if(rating.getPharmacist() != null)
+            return "Farmaceut";
+        else return "Apoteka";
     }
 }
